requirements.txt
Django>=4.2
djangorestframework
psycopg[binary]
gunicorn
Pillow
python-telegram-bot==20.6
